refactored tools to generalize region and hex brushes. Will be necessary going forwards
