+++ MultiHex/core.py
@@ -1,5 +1,4 @@
-from PyQt5.QtWidgets import QGraphicsScene, QGraphicsDropShadowEffect, QGraphicsItem, QGraphicsPolygonItem
-from PyQt5 import QtGui, QtCore
+from PyQt5 import QtCore
 
 try:
     from numpy import sqrt, atan, pi, floor, cos, sin
@@ -36,129 +35,6 @@ def is_number(object):
     except TypeError:
         return(False)
 
-class basic_tool:
-    """
-    Prototype a basic tool 
-    """
-    def __init__(self, parent=None):
-        pass
-    def press(self,event):
-        """
-        Called when the right mouse button is depressed 
-
-        @param event 
-        """
-        pass
-    def activate(self, event):
-        """
-        This is called when the right mouse button is released from a localized click. 
-
-        @param event - location of release
-        """
-        pass
-    def hold(self,event ):
-        """
-        Called continuously while the right mouse button is moved and depressed 
-
-        @param event - current mouse location
-        @param setp  - vector pointing from last called location to @place
-        """
-        pass
-    def select(self, event ):
-        """
-        Left click released event, used to select something
-
-        @param event - Qt event object. has where the mouse is
-        """
-        pass
-    def move(self, event):
-        """
-        Called continuously while the mouse is in the widget
-
-        @param place - where the mouse is 
-        """
-        pass
-    def drop(self):
-        """
-        Called when this tool is being replaced. Cleans up anything it has drawn and should get rid of (like, selection circles)
-        """
-        pass
-    def toggle_mode(self, force=None):
-        pass
-
-class clicker_control(QGraphicsScene):
-    """
-    Manages the mouse interface for to the canvas.
-    """
-    def __init__(self, parent=None, master=None):
-        QGraphicsScene.__init__(self, parent)
-
-        self._active = None
-        self._held = False
-        
-        self.master = master
-
-        self._alt_held = False
-
-    def keyPressEvent(self, event):
-        event.accept()
-        if event.key() == QtCore.Qt.Key_Alt:
-            self._alt_held = True
-    def keyReleaseEvent(self, event):
-        event.accept()
-        if event.key() == QtCore.Qt.Key_Alt:
-            self._alt_held = False
-
-    def mousePressEvent(self, event):
-        """
-        Called whenever the mouse is pressed within its bounds (The drawspace)
-        """
-        if event.button()==QtCore.Qt.RightButton: # or (event.button()==QtCore.Qt.LeftButton and self._alt_held):
-            event.accept() # accept the event
-            self._held = True # say that the mouse is being held 
-            self._active.press( event )
-
-    def mouseReleaseEvent( self, event):
-        """
-        Called when a mouse button is released 
-        """
-        if event.button()==QtCore.Qt.RightButton: # or (event.button()==QtCore.Qt.LeftButton and self._alt_held):
-            # usually a brush event 
-            event.accept()
-            self._held = False
-            self._active.activate(event)
-
-        elif event.button()==QtCore.Qt.LeftButton:
-            # usually a selection event
-            event.accept()
-            self._active.select( event )
-
-   #mouseMoveEvent 
-    def mouseMoveEvent(self,event):
-        """
-        called continuously as the mouse is moved within the graphics space. The "held" boolean is used to distinguish between regular moves and click-drags 
-        """
-        event.accept()
-        if self._held:
-            self._active.hold( event )
- 
-        self._active.move( event )
-
-    # in c++ these could've been templates and that would be really cool 
-    def to_hex(self):
-        """
-        We need to switch over to calling the writer control, and have the selector clean itself up. These two cleaners are used to git rid of any drawn selection outlines 
-        """
-        self._active.drop()
-        self._active = self.master.writer_control
-
-    def to_region(self):
-        """
-        same...
-        """
-        self._active.drop()
-        self._active = self.master.region_control
- 
 class Point:
     """
     A vector in 2D Cartesian space.
